requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
bcrypt==3.1.7
blessings==1.7
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cssselect2==0.3.0
defusedxml==0.6.0
docopt==0.6.2
fabric==2.4.0
fabric2==2.4.0
idna==2.8
importlib-metadata==0.18
invoke==1.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
numpy==1.16.1
packaging==19.0
pandas==1.0.1
paramiko==2.6.0
Pillow==7.0.0
pipreqs==0.4.9
pluggy==0.12.0
pycparser==2.19
pylint==2.2.2
PyNaCl==1.3.0
pysftp==0.2.9
pytest==5.0.0
python-barcode==0.11.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rope==0.14.0
scipy==1.3.0
six==1.11.0
spur==0.3.21
TBB==0.1
tinycss2==1.0.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
xlrd==1.2.0
yarg==0.1.9
zipp==0.5.1
